+++ src/app/cart/models/cart-item.interface.ts
@@ -3,4 +3,6 @@ import { IProductModel } from '../../products/models/product-model.interface';
 export interface ICartItem {
   item: IProductModel;
   count: number;
+  name: string;
+  price: number;
 }